+++ python/GafferSceneTest/TweakPlugTest.py
@@ -84,7 +84,7 @@ class TweakPlugTest( GafferSceneTest.SceneTestCase ) :
 		tweaks.addChild( GafferScene.TweakPlug( "b", 10.0, GafferScene.TweakPlug.Mode.Multiply ) )
 
 		parameters = IECore.CompoundData( { "a" : 0.0, "b" : 2.0 } )
-		tweaks.applyTweaks( parameters )
+		self.assertTrue( tweaks.applyTweaks( parameters ) )
 		self.assertEqual( parameters, IECore.CompoundData( { "a" : 1.0, "b" : 20.0 } ) )
 
 	def testTweakNetwork( self ) :
@@ -103,7 +103,7 @@ class TweakPlugTest( GafferSceneTest.SceneTestCase ) :
 		tweaks.addChild( GafferScene.TweakPlug( "surface.Kd", 0.5, GafferScene.TweakPlug.Mode.Multiply ) )
 		tweaks.addChild( GafferScene.TweakPlug( "Kd", 0.25, GafferScene.TweakPlug.Mode.Add ) )
 
-		tweaks.applyTweaks( network )
+		self.assertTrue( tweaks.applyTweaks( network ) )
 
 		self.assertEqual( network.getShader( "texture" ).parameters["sscale"].value, 11.0 )
 		self.assertEqual( network.getShader( "surface" ).parameters["Kd"].value, 0.75 )
@@ -142,11 +142,11 @@ class TweakPlugTest( GafferSceneTest.SceneTestCase ) :
 		self.assertEqual( d, IECore.CompoundData() )
 
 		d = IECore.CompoundData()
-		p.applyTweaks( d, missingMode = GafferScene.TweakPlug.MissingMode.Ignore )
+		self.assertFalse( p.applyTweaks( d, missingMode = GafferScene.TweakPlug.MissingMode.Ignore ) )
 		self.assertEqual( d, IECore.CompoundData() )
 
 		d = IECore.CompoundData()
-		p.applyTweaks( d, missingMode = GafferScene.TweakPlug.MissingMode.IgnoreOrReplace )
+		self.assertTrue( p.applyTweaks( d, missingMode = GafferScene.TweakPlug.MissingMode.IgnoreOrReplace ) )
 		self.assertEqual( d, IECore.CompoundData( { "test" : 0.5 } ) )
 
 		d = IECore.CompoundData()
@@ -159,7 +159,7 @@ class TweakPlugTest( GafferSceneTest.SceneTestCase ) :
 			p.applyTweaks( d, missingMode = GafferScene.TweakPlug.MissingMode.IgnoreOrReplace )
 		self.assertEqual( d, IECore.CompoundData() )
 
-		p.applyTweaks( d, missingMode = GafferScene.TweakPlug.MissingMode.Ignore )
+		self.assertFalse( p.applyTweaks( d, missingMode = GafferScene.TweakPlug.MissingMode.Ignore ) )
 		self.assertEqual( d, IECore.CompoundData() )
 
 	def testTweaksPlug( self ) :
@@ -199,7 +199,7 @@ class TweakPlugTest( GafferSceneTest.SceneTestCase ) :
 			p.applyTweaks( networkCopy, missingMode = GafferScene.TweakPlug.MissingMode.Error )
 
 		self.assertEqual( networkCopy, network )
-		p.applyTweaks( networkCopy, missingMode = GafferScene.TweakPlug.MissingMode.Ignore )
+		self.assertFalse( p.applyTweaks( networkCopy, missingMode = GafferScene.TweakPlug.MissingMode.Ignore ) )
 		self.assertEqual( networkCopy, network )
 
 		p["t"]["name"].setValue( "missingShader.parameterName" )
@@ -211,8 +211,80 @@ class TweakPlugTest( GafferSceneTest.SceneTestCase ) :
 			p.applyTweaks( networkCopy, missingMode = GafferScene.TweakPlug.MissingMode.Error )
 
 		self.assertEqual( networkCopy, network )
-		p.applyTweaks( networkCopy, missingMode = GafferScene.TweakPlug.MissingMode.Ignore )
+		self.assertFalse( p.applyTweaks( networkCopy, missingMode = GafferScene.TweakPlug.MissingMode.Ignore ) )
 		self.assertEqual( networkCopy, network )
+
+	def testApplyReturnValues( self ) :
+
+		parameters = IECore.CompoundData( { "a" : 0.0, "b" : 2.0 } )
+
+		tweaks = GafferScene.TweaksPlug()
+
+		# Test none to apply
+
+		self.assertFalse( tweaks.applyTweaks( parameters ) )
+
+		# Test none enabled
+
+		tweaks.addChild( GafferScene.TweakPlug( "a", 1.0, GafferScene.TweakPlug.Mode.Replace, False ) )
+		tweaks.addChild( GafferScene.TweakPlug( "b", 10.0, GafferScene.TweakPlug.Mode.Multiply, False ) )
+
+		tweakedParameters = parameters.copy()
+		self.assertFalse( tweaks.applyTweaks( parameters ) )
+		self.assertEqual( tweakedParameters, parameters )
+
+		# Test enabled
+
+		tweaks[0]["enabled"].setValue( True )
+		tweaks[1]["enabled"].setValue( True )
+
+		self.assertTrue( tweaks.applyTweaks( parameters ) )
+
+		# Test non-matching
+
+		altParameters = IECore.CompoundData( { "c" : 0.0, "d" : 2.0 } )
+		tweakedAltParameters = altParameters.copy()
+		self.assertFalse( tweaks.applyTweaks( tweakedAltParameters, missingMode = GafferScene.TweakPlug.MissingMode.Ignore ) )
+		self.assertEqual( tweakedAltParameters, altParameters )
+
+		# Test empty names
+
+		tweaks[0]["name"].setValue( "" )
+		tweaks[1]["name"].setValue( "" )
+
+		tweakedParameters = parameters.copy()
+		self.assertFalse( tweaks.applyTweaks( parameters ) )
+		self.assertEqual( tweakedParameters, parameters )
+
+	def testApplyReturnValuesNetworkEdits( self ) :
+
+		network = IECoreScene.ShaderNetwork(
+			shaders = {
+				"surface" : IECoreScene.Shader( "lambert", "surface", { "c" : imath.Color3f( 1.0 ) } )
+			},
+			output = "surface"
+		)
+
+		textureShader = GafferSceneTest.TestShader( "texture" )
+
+		tweaks = GafferScene.TweaksPlug()
+
+		tweaks.addChild( GafferScene.TweakPlug( "c", Gaffer.Color3fPlug(), GafferScene.TweakPlug.Mode.Replace, False ) )
+		tweaks[0]["value"].setInput( textureShader["out"] )
+
+		# Test none to apply
+
+		tweakedNetwork = network.copy()
+		self.assertFalse( tweaks.applyTweaks( tweakedNetwork ) )
+		self.assertEqual( tweakedNetwork, network )
+
+		# Test enabled
+
+		tweaks[0]["enabled"].setValue( True )
+
+		tweakedNetwork = network.copy()
+		self.assertTrue( tweaks.applyTweaks( tweakedNetwork ) )
+		self.assertEqual( tweakedNetwork.inputConnections( "surface" ), [ ( ( "texture", "" ), ( "surface", "c" ) ) ] )
 
 if __name__ == "__main__":
 	unittest.main()